# new-api-doc.yml
openapi: 3.0.4
info:
  title: API - LPM UMC
  description: >
    Dokumentasi REST API Sistem Lembaga Penjaminan Mutu (LPM) Universitas Muhammadiyah Cirebon.
    API ini digunakan untuk autentikasi, monitoring & evaluasi (Monev),
    penilaian, survei, serta rencana tindak lanjut.
  version: 1.0.0
servers:
  - url: http://localhost:3001/api/{lang}
    description: Server development
    variables:
      lang:
        enum: [id, en]
        default: id
  - url: http://lpm.umc.ac.id/api/{lang}
    description: Server production
    variables:
      lang:
        enum: [id, en]
        default: id
paths:
...
  /auth/google:
    get:
      tags: [Auth]
      summary: redirect ke google oauth
      description: >
        endpoint untuk memulai proses login menggunakan google oauth 2.0.
        server akan menghasilkan url oauth dan me-redirect user ke halaman login google.
      responses:
        "302":
          description: redirect berhasil ke google oauth consent screen
          headers:
            Location:
              description: URL Google OAuth
              schema:
                type: string
              example: https://accounts.google.com/o/oauth2/v2/auth?...
        "500":
          description: Gagal menginisialisasi OAuth Google
...
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout user
      security:
        - bearerAuth: []
      description: >
        Mengakhiri sesi login user dan menghapus token autentikasi.
      responses:
        "204":
          description: Logout berhasil (No content)
        "401":
          description: Invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthenticatedError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"


  # =====================================================
  # Kalender
  # =====================================================
...
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UnauthenticatedError:
      type: object
      properties:
        errors:
          type: string
          example: Unauthenticated
...
    InternalServerError:
      description: Terjadi kesalahan di server internal
      properties:
        errors:
          type: string
          example: Internal server error
...